+++ R/plotMethods.R
@@ -34,7 +34,9 @@ plot.gsmarpred <- function(x, ..., nt, mix_weights=TRUE, add_grid=TRUE) {
   old_par <- par(no.readonly=TRUE) # Save old settings
   on.exit(par(old_par)) # Restore the settings before quitting
   if(mix_weights) {
-    par(mfrow=c(2, 1), mar=c(2.6, 2.6, 2.6, 2.6))
+    par(mfrow=c(2, 1), mar=c(2.1, 2.6, 2.1, 1.1))
+  } else {
+    par(mfrow=c(1, 1), mar=c(2.6, 2.6, 2.1, 1.1))
   }
 
   if(missing(nt)) {
@@ -111,9 +113,10 @@ plot.gsmarpred <- function(x, ..., nt, mix_weights=TRUE, add_grid=TRUE) {
     mix_ts <- ts(mixing_weights[(n_mix - nt + 1):n_mix,], start=time(data)[n_obs - nt + 1],
                  frequency=frequency(data))
     mix_pred_ts <- ts(rbind(mix_ts[nrow(mix_ts),], gsmarpred$mix_pred), start=time(data)[n_obs], frequency=frequency(data))
-    ts.plot(mix_ts, mix_pred_ts, gpars=list(col=c(colpal_mw2, colpal_mw), ylim=c(0, 1), lty=c(rep(1, M), rep(2, M))))
+    ts.plot(mix_ts, mix_pred_ts, gpars=list(col=c(colpal_mw2, colpal_mw), ylim=c(0, 1), lty=c(rep(1, M), rep(2, M))),
+            main="Mixing weights")
     legend("topleft", legend=paste0("regime ", 1:M), bty="n", col=colpal_mw, lty=1, lwd=2,
-           text.font=2, cex=0.65, x.intersp=0.5, y.intersp=1)
+           text.font=2, cex=0.70, x.intersp=0.5, y.intersp=1)
     if(add_grid) grid(...)
 
     # Individual prediction intervals as for the mixing weights
@@ -211,7 +214,7 @@ plot.gsmar <- function(x, ..., include_dens=TRUE) {
   ts.plot(data, gpars=list(main="Time series"))
   ts.plot(ts_mw, gpars=list(main="Mixing weights", ylim=c(0, 1), col=colpal_mw, lty=2))
   legend("topleft", legend=paste0("regime ", 1:M), bty="n", col=colpal_mw, lty=1, lwd=2,
-         text.font=2, cex=0.65, x.intersp=0.5, y.intersp=1)
+         text.font=2, cex=0.75, x.intersp=0.5, y.intersp=1)
 
   # Plot kernel density estimate of the data with the model implied density
   if(include_dens) {