+++ R/quantileResiduals.R
@@ -8,8 +8,6 @@
 #' @return Returns a \eqn{(Tx1)} numeric vector containing the quantile residuals of the specified GMAR, StMAR or G-StMAR model.
 #' @details Numerical integration is employed if the quantile residuals cannot be obtained analytically with the
 #'  hypergeometric function using the package 'gsl'.
-#' @section Suggested packages:
-#'   Install the suggested package "gsl" for faster evaluation of the quantile residuals for the StMAR and G-StMAR models.
 #' @references
 #'  \itemize{
 #'    \item Galbraith, R., Galbraith, J. 1974. On the inverses of some patterned matrices arising
@@ -38,8 +36,6 @@ quantile_residuals_int <- function(data, p, M, params, model=c("GMAR", "StMAR", 
 #'
 #' @inheritParams quantile_residuals_int
 #' @inherit quantile_residuals_int return details references
-#' @section Suggested packages:
-#'   Install the suggested package "gsl" for faster evaluation of the quantile residuals of StMAR and G-StMAR models.
 #' @examples
 #' # StMAR model
 #' params43 <- c(0.09, 1.31, -0.46, 0.33, -0.23, 0.04, 0.01, 1.15,